Move CurrencySettings into new "Asset Settings" Scene
