+++ src/components/Main.tsx
@@ -38,6 +38,7 @@ import { HeaderTitle } from './navigation/HeaderTitle'
 import { ParamHeaderTitle } from './navigation/ParamHeaderTitle'
 import { SideMenuButton } from './navigation/SideMenuButton'
 import { TransactionDetailsTitle } from './navigation/TransactionDetailsTitle'
+import { AssetSettingsScene as AssetSettingsSceneComponent } from './scenes/AssetSettingsScene'
 import { ChangeMiningFeeScene as ChangeMiningFeeSceneComponent } from './scenes/ChangeMiningFeeScene'
 import { ChangeMiningFeeScene2 as ChangeMiningFeeScene2Component } from './scenes/ChangeMiningFeeScene2'
 import { ChangePasswordScene as ChangePasswordSceneComponent } from './scenes/ChangePasswordScene'
@@ -144,6 +145,7 @@ const CryptoExchangeQuoteProcessingScene = ifLoggedIn(CryptoExchangeQuoteProcess
 const CryptoExchangeScene = ifLoggedIn(CryptoExchangeSceneComponent)
 const CryptoExchangeSuccessScene = ifLoggedIn(CryptoExchangeSuccessSceneComponent)
 const CurrencyNotificationScene = ifLoggedIn(CurrencyNotificationSceneComponent)
+const AssetSettingsScene = ifLoggedIn(AssetSettingsSceneComponent)
 const CurrencySettingsScene = ifLoggedIn(CurrencySettingsSceneComponent)
 const DefaultFiatSettingScene = ifLoggedIn(DefaultFiatSettingSceneComponent)
 const EdgeLoginScene = ifLoggedIn(EdgeLoginSceneComponent)
@@ -422,6 +424,14 @@ const EdgeAppStack = () => {
         component={CurrencyNotificationScene}
         options={{
           headerTitle: props => <CurrencySettingsTitle />,
+          headerRight: () => null
+        }}
+      />
+      <Stack.Screen
+        name="assetSettings"
+        component={AssetSettingsScene}
+        options={{
+          title: lstrings.settings_asset_settings,
           headerRight: () => null
         }}
       />
