+++ src/locales/en_US.ts
@@ -390,6 +390,7 @@ const strings = {
   invalid_custom_fee: 'Minimum custom fee is',
   missing_username: '<No username>',
   settings_account_title_cap: 'Account',
+  settings_asset_settings: 'Asset Settings',
   settings_button_change_password: 'Change Password',
   settings_developer_mode: 'Developer Mode',
   settings_verbose_logging: 'Verbose Logging',
